Add VehicleFormAndInputs component with responsive design and image handling
